spatest/requirements.txt
Django
psycopg2
whitenoise
python-dotenv
gunicorn
Pillow
